Core functions

Core operations

Types

The computation graph as a whole is stored in the Graph type. Individual nodes in the graph are referred to as operations, stored in the Operation type. Operations can produce one or more outputs, represented as the Tensor type.

TensorFlow.Tensor — Type.

Represents the output of an operation in the computation graph

source

TensorFlow.Operation — Type.

An operation in the computation graph.

source

TensorFlow.Session — Type.

A TensorFlow session.

source

TensorFlow.Graph — Type.

A TensorFlow computation graph

source

Functions

Base.eltype — Function.

eltype(s::Sampleable)

The default element type of a sample. This is the type of elements of the samples generated by the rand method. However, one can provide an array of different element types to store the samples using rand!.

source

TensorFlow.node_name — Function.

node_name(node::AbstractOperation)

Returns the name of a node in the computation graph.

source

TensorFlow.get_collection — Function.

get_collection(g::Graph, name)

Returns a collection attached to the graph g named name

source

TensorFlow.get_def_graph — Function.

get_def_graph()

Returns the default computation graph, an object of type Graph.

See also as_default for setting the default graph

source

Base.run — Function.

run(sess::Session, output, input_dict::Dict)

Compute the result of one of more operations in the computation graph.

source

TensorFlow.get_node_by_name — Function.

get_node_by_name(graph::Graph, name::AbstractString)

Returns an operation by searching for its name in the given graph.

source

TensorFlow.get_shape — Function.

Runs shape inference to return the shape of the tensor produced by the given operation.

Note this runs statically. Use the shape operation to dynamically get the shape of an operation.

source

TensorFlow.get_def — Function.

get_def(n)

Returns the definition of the given operation or graph

source