Summaries

Summary reference

Writing event files

TensorFlow.summary.FileWriter — Type.

FileWriter(logdir; graph=get_def_graph())

The FileWriter type provides a mechanism to create an event file in a given directory and add summaries and events to it. The class updates the file contents asynchronously. This allows a training program to call methods to add data to the file directly from the training loop, without slowing down training.

On construction the summary writer creates a new event file in logdir.

If you pass a Graph to the constructor it is added to the event file.

Arguments:

logdir: A string. Directory where event file will be written.

graph: A Graph object.

source

Summary operations

TensorFlow.summary.summary_ops.scalar — Function.

 scalar_summary(tags, values)

Outputs a Summary protocol buffer with scalar values.

The input tags and values must have the same shape. The generated summary has a summary value for each tag-value pair in tags and values.

source

TensorFlow.summary.summary_ops.histogram — Function.

 histogram_summary(tag, values)

Outputs a Summary protocol buffer with a histogram.

The generated Summary has one summary value containing a histogram for values.

This op reports an InvalidArgument error if any value is not finite.

source

TensorFlow.summary.summary_ops.image — Function.

 image_summary(tag, tensor; max_images=3, bad_color=?)

Outputs a Summary protocol buffer with images.

The summary has up to max_images summary values containing images. The images are built from tensor which must be 4-D with shape [batch_size, height, width, channels] and where channels can be:

1: tensor is interpreted as Grayscale.

3: tensor is interpreted as RGB.

4: tensor is interpreted as RGBA.

The images have the same number of channels as the input tensor. For float input, the values are normalized one image at a time to fit in the range [0, 255]. uint8 values are unchanged. The op uses two different normalization algorithms:

If the input values are all positive, they are rescaled so the largest one is 255.

If any input value is negative, the values are shifted so input value 0.0 is at 127. They are then rescaled so that either the smallest value is 0, or the largest one is 255.

The tag argument is a scalar Tensor of type string. It is used to build the tag of the summary values:

If max_images is 1, the summary value tag is 'tag/image'.

If max_images is greater than 1, the summary value tags are generated sequentially as 'tag/image/0', 'tag/image/1', etc.

The bad_color argument is the color to use in the generated images for non-finite input values. It is a unit8 1-D tensor of length channels. Each element must be in the range [0, 255] (It represents the value of a pixel in the output image). Non-finite values in the input tensor are replaced by this tensor in the output image. The default value is the color red.

source

TensorFlow.summary.summary_ops.merge_all — Function.

merge_all(key=:Summaries)

Merges all summaries collected in the default graph.

Args: key: GraphKey used to collect the summaries. Defaults to :Summaries

Returns: If no summaries were collected, returns nothing. Otherwise returns a scalar Tensor of type String containing the serialized Summary protocol buffer resulting from the merging.

source